# challenge/challenge.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "challenge.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/challenge/alphanumeric": {
      "get": {
        "summary": "Get an alphanumeric challenge",
        "description": "Returns an image with alphanumeric characters to be found, and a prompt if applicable",
        "operationId": "GetAlphanumericChallenge",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/challengeGetAlphanumericResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Challenge"
        ]
      }
    },
    "/challenge/images": {
      "get": {
        "summary": "Get an image challenge",
        "description": "Returns a list of images and a prompt that lets the client know which images we are asking for.",
        "operationId": "GetImagesChallenge",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/challengeGetImagesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "number",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Challenge"
        ]
      }
    },
    "/challenge/questions": {
      "get": {
        "summary": "Get a questions list challenge",
        "description": "Returns a list of questions that are to be answered by the client.",
        "operationId": "GetQuestionsChallenge",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/challengeGetQuestionsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "number",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Challenge"
        ]
      }
    },
    "/solve": {
      "post": {
        "summary": "Solve the session",
        "description": "Returns a success response on pass, or error on failure.",
        "operationId": "SolveSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/challengeSolutionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/challengeSessionSolution"
            }
          }
        ],
        "tags": [
          "Challenge"
        ]
      }
    }
  },
  "definitions": {
    "challengeGetAlphanumericResponse": {
      "type": "object",
      "properties": {
        "session_id": {
          "type": "string"
        },
        "prompt": {
          "type": "string"
        },
        "image": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "challengeGetImagesResponse": {
      "type": "object",
      "properties": {
        "session_id": {
          "type": "string"
        },
        "prompt": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "byte"
          }
        }
      }
    },
    "challengeGetQuestionsResponse": {
      "type": "object",
      "properties": {
        "session_id": {
          "type": "string"
        },
        "questions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "challengeSessionSolution": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "answer": {
          "type": "string"
        }
      }
    },
    "challengeSolutionResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}